feat: 新增土壤湿度(ADC)和BH1750光照(I2C)传感器
传感器节点:
- 新增 soil_moisture 驱动 (ADC1-CH0/PA0, 均值滤波, 0~100%)
- 新增 bh1750 驱动 (I2C1 PB6/PB7, 连续高分辨率, lux)
- SensorData_t: light 扩展为 uint16 支持完整 lux 值
- CAN帧: [5-6]=光照lux, [7]=保留
- Task_Sensor_Acq 同时采集4种数据

网关节点:
- GatewayData_t 新增 soil_moisture + light 字段
- JSON 输出增加 soil 和 light
- OLED 4行轮播显示: 温度/湿度/土壤/光照

+++ common/drivers/can/can_driver.h
@@ -70,8 +70,8 @@ typedef struct {
 typedef struct {
     int16_t  temperature;   /* 温度 ×10 (如 255 = 25.5°C) */
     int16_t  humidity;      /* 湿度 ×10 (如 600 = 60.0%) */
-    uint8_t  soil_moisture; /* 土壤湿度 % (预留) */
-    uint8_t  light;         /* 光照 (预留) */
+    uint8_t  soil_moisture; /* 土壤湿度 0~100% */
+    uint16_t light;         /* 光照强度 (lux) */
 } SensorData_t;
 
 /* ============================================================
